+++ packages/prettier-config/test/index.js
@@ -1,15 +1,11 @@
 /**
- * External dependencies
- */
-import isPlainObject from 'is-plain-obj';
-
-/**
  * Internal dependencies
  */
 import config from '../lib/';
 
 describe( 'prettier config tests', () => {
 	it( 'should be an object', () => {
-		expect( isPlainObject( config ) ).toBeTruthy();
+		expect( config ).not.toBeNull();
+		expect( typeof config ).toBe( 'object' );
 	} );
 } );