+++ packages/blocks/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 11.15.0 (2022-08-24)
+
+### Bug Fix
+
+-   Packages: Replace `is-plain-obj` with `is-plain-object` ([#43511](https://github.com/WordPress/gutenberg/pull/43511)).
+
 ## 11.14.0 (2022-08-10)
 
 ## 11.13.0 (2022-07-27)